Add custom_components demo to the showcase
