+++ examples/showcase/sub/custom_components/Cargo.toml
@@ -4,4 +4,4 @@ version = "0.1.0"
 authors = ["Denis Kolodin <deniskolodin@gmail.com>"]
 
 [dependencies]
-yew = { path = "../.." }
+yew = { path = "../../../.." }
